Functionality implementation WIP
More like RIP
Got a canvas object contain grid objects to paint individually, but there are a ton of concerns regarding implementation, scalability, asyncronous vs syncronous, rendering and re-rendering (draw and re-draw) for zooming and dragging, and just a lot of js horrors. Commiting current progress and putting this on hold.

Experimenting alternative on a separate branch.